+++ Modules/booking/Model/BookingInvoice.php
@@ -133,8 +133,8 @@ class BookingInvoice extends InvoiceModel {
         $modelClient = new ClClient();
         $LABpricingid = $modelClient->getPricingID($id_space, $id_resp);
         Configuration::getLogger()->debug('[invoice][booking] get pricing for client', ['client' => $id_resp, 'id' => $LABpricingid]);
-        $modelResouces = new ResourceInfo();
-        $resources = $modelResouces->getBySpace($id_space);
+        $modelResources = new ResourceInfo();
+        $resources = $modelResources->getBySpace($id_space);
 
         // get the pricing
         $timePrices = $this->getUnitTimePricesForEachResource($id_space, $resources, $LABpricingid, $id_resp);
@@ -151,19 +151,12 @@ class BookingInvoice extends InvoiceModel {
         foreach ($resources as $res) {
             $reservations = $modelCal->getUnpricedReservations($id_space, $beginPeriod, $endPeriod, $res["id"], $id_resp);
 
-            // get list of quantities
-            $calQuantities = $bkCalQuantitiesModel->getByResource($id_space, $res["id"]);
-            $calQuantities = ($calQuantities != null) ? $calQuantities : [];
-
-            // tell if there's an invoicing unit for this resource amongst quantities and get its ID
-            $isInvoicingUnit = false;
-            $calQuantityId = "";
-            foreach ($calQuantities as $calQuantity) {
-                if ($calQuantity["is_invoicing_unit"] && intval($calQuantity["is_invoicing_unit"]) === 1) {
-                    $calQuantityId = $calQuantity["id"];
-                    $isInvoicingUnit = true;
-                }
-            }
+            // get list of quantities sorted by deleted (ASC), then by id (desc)
+            $allCalQuantities = $bkCalQuantitiesModel->getByResource($id_space, $res["id"], include_deleted:true, sort:true);
+            // get invoicable ones
+            $invoicableCalQtes = array_filter($allCalQuantities, function($calQte) {
+                return $calQte["is_invoicing_unit"] == 1;
+            });
 
             // get all packages
             $userPackages = array();
@@ -187,8 +180,36 @@ class BookingInvoice extends InvoiceModel {
             // $userDetails = [];
 
             $totalQte = 0; // $totalQte = total number of items booked
+            
 
             foreach ($reservations as $reservation) {
+                $resaIsInvoicingUnit = false;
+                $resaQuantityId = "";
+                /*
+                 * For a reservation with invoicing units:
+                 * if there is a non deleted invoicing unit for this resource and it is used in this reservation, this is the one which will be used: it replaces the deleted one.
+                 * * if a deleted invoicing unit is used and there is no current invoicing unit used in this reservation, then the deleted one will be used
+                 * if there are more than 1 deleted invoicing unit used in this reservation: use the last created one (max id)
+                 */
+                if ($reservation["quantities"] && $reservation["quantities"] != null) {
+                    // Genrate an array of quantity's ids used in this reservation
+                    $resaQtes = explode(";", $reservation["quantities"]);
+                    array_pop($resaQtes);
+                    $resaQteIds = array_map(function($qte) {
+                        return explode("=", $qte)[0];
+                    }, $resaQtes);
+
+
+                    // Is one of them amongst the deleted invoicable quantities?
+                    // If there are more than one, raises an exception
+                    foreach ($invoicableCalQtes as $invoicableQte) {
+                        if (in_array($invoicableQte['id'], $resaQteIds)) {
+                            $resaQuantityId = $invoicableQte['id'];
+                            $resaIsInvoicingUnit = true;
+                            break;
+                        }
+                    }
+                }
 
                 // count: day night we, packages
                 if ($reservation["package_id"] > 0) {
@@ -214,11 +235,15 @@ class BookingInvoice extends InvoiceModel {
                     $resaDayNightWe = $slots['hours'];
                     Configuration::getLogger()->debug('[invoice][booking] night and week ends', ['resource' => $res['id'], 'count' => $resaDayNightWe]);
 
-                    if ($isInvoicingUnit) {
+                    if ($resaIsInvoicingUnit) {
                         if ($reservation["quantities"] && $reservation["quantities"] != null) {
+                            
+
+
                             // varchar formatted like "$mandatory=$quantity;" in bk_calendar_entry
                             // get number of resources booked
-                            $strToFind = strval($calQuantityId) . "=";
+                            $strToFind = strval($resaQuantityId) . "=";
+                            
                             $lastPos = 0;
                             $positions = array();
                             while(($lastPos = strpos($reservation["quantities"], $strToFind, $lastPos))!==false) {
@@ -227,6 +252,7 @@ class BookingInvoice extends InvoiceModel {
                             }
                             $foundStr = substr($reservation["quantities"], $positions[0]);
                             $qte = intval(explode("=", $foundStr)[1]);
+
                         } else {
                             $qte = 0;
                         }
@@ -252,7 +278,6 @@ class BookingInvoice extends InvoiceModel {
                             'user' => $reservation['recipient_id']
                         ];
                         */
-                        
                     } else {
                         $userTime["nb_hours_day"] += $resaDayNightWe["nb_hours_day"];
                         $userTime["nb_hours_night"] += $resaDayNightWe["nb_hours_night"];
@@ -269,7 +294,7 @@ class BookingInvoice extends InvoiceModel {
                             'user' => $reservation['recipient_id']
                         ];
                         */
-                    }                    
+                    }              
                 }
 
                 $modelCal->setReservationInvoice($id_space, $reservation["id"], $invoice_id);
@@ -552,7 +577,7 @@ class BookingInvoice extends InvoiceModel {
         $modelPriceOwner = new BkOwnerPrice();
         foreach ($resources as $resource) {
             // get the packages prices
-            $packages = $modelPackage->getByResource($id_space ,$resource["id"]);
+            $packages = $modelPackage->getByResource($id_space ,$resource["id"], include_deleted:true);
 
             $pricesPackages = array();
             for ($i = 0; $i < count($packages); $i++) {